Extract some generic bndlib code into a common module
(cherry picked from commit c478f659b626587529ee158d6011ac3d43c5151a)
